requirements.txt
PyQt5==5.15.10
PyYAML==6.0.1
openpyxl==3.1.2
pyinstaller==6.3.0
